fix(#1488): fail-closed on clustering key named 'inclusive' in delta schema
The delta envelope's range-bound structs (__range_start/__range_end)
append a hardcoded 'inclusive: bool' sub-field after one field per
clustering column. A table with a CLUSTERING column literally named
'inclusive' produced two 'inclusive' fields in that struct -> malformed
Arrow schema that failed LATE at StructArray::try_new (after the temp
output file was created) instead of fail-before-writing.

Extend the collision check in derive_delta_schema to reject any
clustering-key column named 'inclusive' (the range-bound reserved
sub-field), raising DeltaSchemaError::ColumnCollision at derivation time.
Since delta_parquet's writer derives the schema before building any
arrays, this covers the mirror builder path too. The sub-field is
hardcoded (not namespaced by envelope_prefix), so the collision fires
regardless of prefix. Partition-key and regular columns named
'inclusive' are unaffected (they never enter the range-bound struct).

TDD: added clustering_key_named_inclusive_collision (+ custom-prefix and
regular-column-allowed variants); the collision test fails on main
(returns Ok with a duplicate-field schema) and passes after the fix.

Extracted the test module into delta_schema_tests.rs (epic #1135) to
keep delta_schema.rs under the campsite-rule threshold.
